feat: disable edge lambda as its not really used.
The lambda@edge was meant to use a dyanmodb table to validate api keys this is no longer a requirement so it can be safely removed
