--- packages/lambda-api-tracker/README.md
@@ -1,1 +1,0 @@
-# @basemaps/lambda-api-tracker
