--- packages/lambda-api-tracker/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@basemaps/lambda-api-tracker",
-  "version": "6.1.0",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/linz/basemaps.git",
-    "directory": "packages/lambda-api-tracker"
-  },
-  "author": {
-    "name": "Land Information New Zealand",
-    "url": "https://linz.govt.nz",
-    "organization": true
-  },
-  "license": "MIT",
-  "dependencies": {
-    "@basemaps/geo": "^6.0.0",
-    "@basemaps/lambda": "^6.1.0",
-    "@basemaps/shared": "^6.1.0"
-  },
-  "scripts": {
-    "test": "ospec --globs 'build/**/*.test.js'",
-    "bundle": "../../scripts/bundle.js package.json"
-  },
-  "devDependencies": {
-    "@types/aws-lambda": "^8.10.75",
-    "@types/node": "^14.11.2"
-  },
-  "bundle": {
-    "entry": "src/index.ts",
-    "outdir": "dist/",
-    "external": [
-      "aws-sdk",
-      "pino-pretty"
-    ]
-  }
-}
